@@ -2,8 +2,10 @@ import argparse
 
 
 class ScaleneArguments(argparse.Namespace):
+    """Encapsulates all arguments and default values for Scalene."""
+
     def __init__(self) -> None:
-        super(ScaleneArguments, self).__init__()
+        super().__init__()
         self.cpu_only = False
         self.cpu_percent_threshold = 1
         # mean seconds between interrupts for CPU sampling.
@@ -34,7 +36,7 @@ class ScaleneArguments(argparse.Namespace):
         self.reduced_profile = False
         # do we use virtual time or wallclock time (capturing system time and blocking)?
         self.use_virtual_time = False
-        self.memory_leak_detector = False  # experimental
+        self.memory_leak_detector = True  # experimental
         self.web = True
         self.port = 8088
         self.cli = False
