@@ -1,20 +1,20 @@
-;(function() {
+(function () {
   function Tablesort(el, options) {
     if (!(this instanceof Tablesort)) return new Tablesort(el, options);
 
-    if (!el || el.tagName !== 'TABLE') {
-      throw new Error('Element must be a table');
+    if (!el || el.tagName !== "TABLE") {
+      throw new Error("Element must be a table");
     }
     this.init(el, options || {});
   }
 
   var sortOptions = [];
 
-  var createEvent = function(name) {
+  var createEvent = function (name) {
     var evt;
 
-    if (!window.CustomEvent || typeof window.CustomEvent !== 'function') {
-      evt = document.createEvent('CustomEvent');
+    if (!window.CustomEvent || typeof window.CustomEvent !== "function") {
+      evt = document.createEvent("CustomEvent");
       evt.initCustomEvent(name, false, false, undefined);
     } else {
       evt = new CustomEvent(name);
@@ -23,12 +23,17 @@
     return evt;
   };
 
-  var getInnerText = function(el,options) {
-    return el.getAttribute(options.sortAttribute || 'data-sort') || el.textContent || el.innerText || '';
+  var getInnerText = function (el, options) {
+    return (
+      el.getAttribute(options.sortAttribute || "data-sort") ||
+      el.textContent ||
+      el.innerText ||
+      ""
+    );
   };
 
   // Default sort method if no better sort method is found
-  var caseInsensitiveSort = function(a, b) {
+  var caseInsensitiveSort = function (a, b) {
     a = a.trim().toLowerCase();
     b = b.trim().toLowerCase();
 
@@ -38,17 +43,17 @@
     return -1;
   };
 
-  var getCellByKey = function(cells, key) {
-    return [].slice.call(cells).find(function(cell) {
-      return cell.getAttribute('data-sort-column-key') === key;
+  var getCellByKey = function (cells, key) {
+    return [].slice.call(cells).find(function (cell) {
+      return cell.getAttribute("data-sort-column-key") === key;
     });
   };
 
   // Stable sort function
   // If two elements are equal under the original sort function,
   // then there relative order is reversed
-  var stabilize = function(sort, antiStabilize) {
-    return function(a, b) {
+  var stabilize = function (sort, antiStabilize) {
+    return function (a, b) {
       var unstableResult = sort(a.td, b.td);
 
       if (unstableResult === 0) {
@@ -60,26 +65,25 @@
     };
   };
 
-  Tablesort.extend = function(name, pattern, sort) {
-    if (typeof pattern !== 'function' || typeof sort !== 'function') {
-      throw new Error('Pattern and sort must be a function');
+  Tablesort.extend = function (name, pattern, sort) {
+    if (typeof pattern !== "function" || typeof sort !== "function") {
+      throw new Error("Pattern and sort must be a function");
     }
 
     sortOptions.push({
       name: name,
       pattern: pattern,
-      sort: sort
+      sort: sort,
     });
   };
 
   Tablesort.prototype = {
-
-    init: function(el, options) {
+    init: function (el, options) {
       var that = this,
-          firstRow,
-          defaultSort,
-          i,
-          cell;
+        firstRow,
+        defaultSort,
+        i,
+        cell;
 
       that.table = el;
       that.thead = false;
@@ -88,7 +92,7 @@
       if (el.rows && el.rows.length > 0) {
         if (el.tHead && el.tHead.rows.length > 0) {
           for (i = 0; i < el.tHead.rows.length; i++) {
-            if (el.tHead.rows[i].getAttribute('data-sort-method') === 'thead') {
+            if (el.tHead.rows[i].getAttribute("data-sort-method") === "thead") {
               firstRow = el.tHead.rows[i];
               break;
             }
@@ -104,9 +108,9 @@
 
       if (!firstRow) return;
 
-      var onClick = function() {
+      var onClick = function () {
         if (that.current && that.current !== this) {
-          that.current.removeAttribute('aria-sort');
+          that.current.removeAttribute("aria-sort");
         }
 
         that.current = this;
@@ -116,12 +120,12 @@
       // Assume first row is the header and attach a click handler to each.
       for (i = 0; i < firstRow.cells.length; i++) {
         cell = firstRow.cells[i];
-        cell.setAttribute('role','columnheader');
-        if (cell.getAttribute('data-sort-method') !== 'none') {
+        cell.setAttribute("role", "columnheader");
+        if (cell.getAttribute("data-sort-method") !== "none") {
           cell.tabindex = 0;
-          cell.addEventListener('click', onClick, false);
+          cell.addEventListener("click", onClick, false);
 
-          if (cell.getAttribute('data-sort-default') !== null) {
+          if (cell.getAttribute("data-sort-default") !== null) {
             defaultSort = cell;
           }
         }
@@ -133,30 +137,30 @@
       }
     },
 
-    sortTable: function(header, update) {
+    sortTable: function (header, update) {
       var that = this,
-          columnKey = header.getAttribute('data-sort-column-key'),
-          column = header.cellIndex,
-          sortFunction = caseInsensitiveSort,
-          item = '',
-          items = [],
-          i = that.thead ? 0 : 1,
-          sortMethod = header.getAttribute('data-sort-method'),
-          sortOrder = header.getAttribute('aria-sort');
+        columnKey = header.getAttribute("data-sort-column-key"),
+        column = header.cellIndex,
+        sortFunction = caseInsensitiveSort,
+        item = "",
+        items = [],
+        i = that.thead ? 0 : 1,
+        sortMethod = header.getAttribute("data-sort-method"),
+        sortOrder = header.getAttribute("aria-sort");
 
-      that.table.dispatchEvent(createEvent('beforeSort'));
+      that.table.dispatchEvent(createEvent("beforeSort"));
 
       // If updating an existing sort, direction should remain unchanged.
       if (!update) {
-        if (sortOrder === 'ascending') {
-          sortOrder = 'descending';
-        } else if (sortOrder === 'descending') {
-          sortOrder = 'ascending';
+        if (sortOrder === "ascending") {
+          sortOrder = "descending";
+        } else if (sortOrder === "descending") {
+          sortOrder = "ascending";
         } else {
-          sortOrder = that.options.descending ? 'descending' : 'ascending';
+          sortOrder = that.options.descending ? "descending" : "ascending";
         }
 
-        header.setAttribute('aria-sort', sortOrder);
+        header.setAttribute("aria-sort", sortOrder);
       }
 
       if (that.table.rows.length < 2) return;
@@ -165,14 +169,14 @@
       if (!sortMethod) {
         var cell;
         while (items.length < 3 && i < that.table.tBodies[0].rows.length) {
-          if(columnKey) {
+          if (columnKey) {
             cell = getCellByKey(that.table.tBodies[0].rows[i].cells, columnKey);
           } else {
             cell = that.table.tBodies[0].rows[i].cells[column];
           }
 
           // Treat missing cells as empty cells
-          item = cell ? getInnerText(cell,that.options) : "";
+          item = cell ? getInnerText(cell, that.options) : "";
 
           item = item.trim();
 
@@ -204,10 +208,10 @@
 
       for (i = 0; i < that.table.tBodies.length; i++) {
         var newRows = [],
-            noSorts = {},
-            j,
-            totalRows = 0,
-            noSortsSoFar = 0;
+          noSorts = {},
+          j,
+          totalRows = 0,
+          noSortsSoFar = 0;
 
         if (that.table.tBodies[i].rows.length < 2) continue;
 
@@ -215,7 +219,7 @@
           var cell;
 
           item = that.table.tBodies[i].rows[j];
-          if (item.getAttribute('data-sort-method') === 'none') {
+          if (item.getAttribute("data-sort-method") === "none") {
             // keep no-sorts in separate list to be able to insert
             // them back at their original position later
             noSorts[totalRows] = item;
@@ -228,8 +232,8 @@
             // Save the index for stable sorting
             newRows.push({
               tr: item,
-              td: cell ? getInnerText(cell,that.options) : '',
-              index: totalRows
+              td: cell ? getInnerText(cell, that.options) : "",
+              index: totalRows,
             });
           }
           totalRows++;
@@ -237,7 +241,7 @@
         // Before we append should we reverse the new array or not?
         // If we reverse, the sort needs to be `anti-stable` so that
         // the double negatives cancel out
-        if (sortOrder === 'descending') {
+        if (sortOrder === "descending") {
           newRows.sort(stabilize(sortFunction, true));
         } else {
           newRows.sort(stabilize(sortFunction, false));
@@ -259,17 +263,17 @@
         }
       }
 
-      that.table.dispatchEvent(createEvent('afterSort'));
+      that.table.dispatchEvent(createEvent("afterSort"));
     },
 
-    refresh: function() {
+    refresh: function () {
       if (this.current !== undefined) {
         this.sortTable(this.current, true);
       }
-    }
+    },
   };
 
-  if (typeof module !== 'undefined' && module.exports) {
+  if (typeof module !== "undefined" && module.exports) {
     module.exports = Tablesort;
   } else {
     window.Tablesort = Tablesort;
