@@ -1,51 +1,53 @@
 import platform
 import re
 import subprocess
+from typing import Tuple
 
 
 class ScaleneAppleGPU:
+    """Wrapper class for Apple integrated GPU statistics."""
+
     def __init__(self) -> None:
         assert platform.system() == "Darwin"
         self.cmd = (
             'DYLD_INSERT_LIBRARIES="" ioreg -r -d 1 -w 0 -c "IOAccelerator"'
         )
-        self.regex = re.compile(r'"Device Utilization %"=(\d+)')
-        self.last_load = 0.0
-        self.sample_count = 0
-        self.sample_interval = 10
-        pass
+        self.regex_util = re.compile(r'"Device Utilization %"=(\d+)')
+        self.regex_inuse = re.compile(r'"In use system memory"=(\d+)')
 
     def has_gpu(self) -> bool:
-        # As far as I am aware, all Macs have had integrated GPUs for some time.
+        """Returns true: as far as I am aware, all Macs have had integrated GPUs for some time."""
         return True
 
     def nvml_reinit(self) -> None:
-        # Here for compatibility with the nvidia wrapper.
-        pass
+        """A NOP, here for compatibility with the nvidia wrapper."""
+        return
 
-    def load(self) -> float:
+    def get_stats(self) -> Tuple[float, float]:
+        """Returns a tuple of (utilization%, memory in use)"""
         if not self.has_gpu():
-            return 0.0
-        self.sample_count += 1
-        if self.sample_count < self.sample_interval:
-            return self.last_load
+            return (0.0, 0.0)
         try:
-            s = subprocess.Popen(self.cmd, shell=True, stdout=subprocess.PIPE)
-            s_return = s.stdout.readlines()
-            for line in s_return:
-                line = line.decode("utf-8")
-                if "Device Utilization %" in line:
-                    util = self.regex.search(line)
-                    if util:
-                        self.last_load = int(util.group(1)) / 100.0
-                        self.sample_count = 0
-                        if self.last_load < 0.15:
-                            self.last_load = 0.0
-                        return self.last_load
-            return 0.0  # Fall-through case
+            in_use = 0.0
+            util = 0.0
+            read_process = subprocess.Popen(
+                self.cmd, shell=True, stdout=subprocess.PIPE
+            )
+            if read_process.stdout is not None:
+                read_process_return = read_process.stdout.readlines()
+                for line in read_process_return:
+                    decoded_line = line.decode("utf-8")
+                    if "In use system memory" in decoded_line:
+                        in_use_re = self.regex_inuse.search(decoded_line)
+                        if in_use_re:
+                            in_use = float(in_use_re.group(1))
+                    if "Device Utilization %" in decoded_line:
+                        util_re = self.regex_util.search(decoded_line)
+                        if util_re:
+                            util = int(util_re.group(1)) / 1000
+                    if util and in_use:
+                        break
+                return (util, in_use)
         except:
-            return 0.0
-
-    def memory_used(self) -> int:
-        """Not yet implemented."""
-        return 0
+            pass
+        return (0.0, 0.0)
