@@ -1,8 +1,13 @@
 import copy
 import linecache
+import random
+import re
+from collections import OrderedDict, defaultdict
+from operator import itemgetter
 from pathlib import Path
 from typing import Any, Callable, Dict, List
 
+from scalene.scalene_leak_analysis import ScaleneLeakAnalysis
 from scalene.scalene_statistics import Filename, LineNumber, ScaleneStatistics
 
 
@@ -14,12 +19,48 @@ class ScaleneJSON:
     # Default threshold for number of mallocs to report a file.
     malloc_threshold = 1  # 100
 
+    # Fraction of the maximum footprint to use as granularity for memory timelines
+    # (used for compression). E.g., 10 => 1/10th of the max.
+    memory_granularity_fraction = 10
+
+    # Maximum number of sparkline samples.
+    max_sparkline_samples = 100
+
     def __init__(self) -> None:
         # where we write profile info
         self.output_file = ""
 
         # if we are on a GPU or not
         self.gpu = False
+
+    def compress_samples(
+        self, uncompressed_samples: List[Any], max_footprint: float
+    ) -> List[Any]:
+        # Compress the samples so that the granularity is at least
+        # a certain fraction of the maximum footprint.
+        samples = []
+        granularity = max_footprint / self.memory_granularity_fraction
+
+        last_mem = 0
+        for (t, mem) in uncompressed_samples:
+            if abs(mem - last_mem) >= granularity:
+                # We're above the granularity.
+                # Force all memory amounts to be positive.
+                mem = max(0, mem)
+                # Add a tiny bit of random noise to force different values (for the GUI).
+                mem += abs(random.gauss(0.01, 0.01))
+                # Now we append it and set the last amount to be the
+                # current footprint.
+                samples.append([t, mem])
+                last_mem = mem
+
+        if len(samples) > self.max_sparkline_samples:
+            # Too many samples. We randomly downsample.
+            samples = sorted(
+                random.sample(samples, self.max_sparkline_samples)
+            )
+
+        return samples
 
     # Profile output methods
     def output_profile_line(
@@ -41,10 +82,10 @@ class ScaleneJSON:
         # Correct for negative CPU sample counts. This can happen
         # because of floating point inaccuracies, since we perform
         # subtraction to compute it.
-        if n_cpu_samples_c < 0:
-            n_cpu_samples_c = 0
+        n_cpu_samples_c = max(0, n_cpu_samples_c)
         n_cpu_samples_python = stats.cpu_samples_python[fname][line_no]
         n_gpu_samples = stats.gpu_samples[fname][line_no]
+        n_gpu_mem_samples = stats.gpu_mem_samples[fname][line_no]
 
         # Compute percentages of CPU time.
         if stats.total_cpu_samples:
@@ -111,9 +152,11 @@ class ScaleneJSON:
         else:
             n_copy_mb_s = 0
 
-        samples = stats.per_line_footprint_samples[fname][line_no] # FIXME
-        #if not any(samples):
-        #    samples = []
+        samples = self.compress_samples(
+            stats.per_line_footprint_samples[fname][line_no],
+            stats.max_footprint,
+        )
+
         return {
             "lineno": line_no,
             "line": linecache.getline(fname, line_no),
@@ -121,6 +164,8 @@ class ScaleneJSON:
             "n_cpu_percent_python": n_cpu_percent_python,
             "n_sys_percent": n_sys_percent,
             "n_gpu_percent": n_gpu_percent,
+            "n_gpu_avg_memory_mb": n_gpu_mem_samples.mean() / 1048576,
+            "n_gpu_peak_memory_mb": n_gpu_mem_samples.peak() / 1048576,
             "n_peak_mb": n_peak_mb,
             "n_growth_mb": n_peak_mb,  # For backwards compatibility
             "n_avg_mb": n_avg_mb,
@@ -134,6 +179,7 @@ class ScaleneJSON:
 
     def output_profiles(
         self,
+        program: Filename,
         stats: ScaleneStatistics,
         pid: int,
         profile_this_code: Callable[[Filename, LineNumber], bool],
@@ -166,7 +212,10 @@ class ScaleneJSON:
             return {}
         growth_rate = 0.0
         if profile_memory:
-            samples = stats.memory_footprint_samples
+            samples = self.compress_samples(
+                stats.memory_footprint_samples, stats.max_footprint
+            )
+
             # Compute growth rate (slope), between 0 and 1.
             if stats.allocation_velocity[1] > 0:
                 growth_rate = (
@@ -178,13 +227,14 @@ class ScaleneJSON:
             samples = []
 
         output: Dict[str, Any] = {
+            "program": program,
             "elapsed_time_sec": stats.elapsed_time,
             "growth_rate": growth_rate,
-            "samples": samples,
             "max_footprint_mb": stats.max_footprint,
             "files": {},
             "gpu": self.gpu,
-            "memory": profile_memory
+            "memory": profile_memory,
+            "samples": samples,
         }
 
         # Build a list of files we will actually report on.
@@ -224,11 +274,43 @@ class ScaleneJSON:
             # If the file was actually a Jupyter (IPython) cell,
             # restore its name, as in "[12]".
             fname_print = fname
-            import re
 
             result = re.match("<ipython-input-([0-9]+)-.*>", fname_print)
             if result:
                 fname_print = Filename("[" + result.group(1) + "]")
+
+            # Leak analysis
+            # First, compute AVERAGE memory consumption.
+            avg_mallocs: Dict[LineNumber, float] = defaultdict(float)
+            for line_no in stats.bytei_map[fname]:
+                n_malloc_mb = stats.memory_aggregate_footprint[fname][line_no]
+                count = stats.memory_malloc_count[fname][line_no]
+                if count:
+                    avg_mallocs[line_no] = n_malloc_mb / count
+                else:
+                    # Setting to n_malloc_mb addresses the edge case where this allocation is the last line executed.
+                    avg_mallocs[line_no] = n_malloc_mb
+
+            avg_mallocs = OrderedDict(
+                sorted(avg_mallocs.items(), key=itemgetter(1), reverse=True)
+            )
+
+            # Now only report potential leaks if the allocation
+            # velocity (growth rate) is above some threshold.
+            leaks = ScaleneLeakAnalysis.compute_leaks(
+                growth_rate, stats, avg_mallocs, fname
+            )
+
+            # Sort in descending order by least likelihood
+            leaks = sorted(leaks, key=itemgetter(1), reverse=True)
+
+            reported_leaks = {}
+
+            for (leak_lineno, leak_likelihood, leak_velocity) in leaks:
+                reported_leaks[str(leak_lineno)] = {
+                    "likelihood": leak_likelihood,
+                    "velocity_mb_s": leak_velocity / stats.elapsed_time,
+                }
 
             # Print header.
             if not stats.total_cpu_samples:
@@ -241,13 +323,14 @@ class ScaleneJSON:
             # Print out the the profile for the source, line by line.
             with open(fname, "r", encoding="utf-8") as source_file:
                 code_lines = source_file.readlines()
-                
+
                 output["files"][fname_print] = {
                     "percent_cpu_time": percent_cpu_time,
                     "lines": [],
+                    "leaks": reported_leaks,
                 }
-                for line_no, line in enumerate(code_lines, start=1):
-                    o = self.output_profile_line(
+                for line_no, _line in enumerate(code_lines, start=1):
+                    profile_line = self.output_profile_line(
                         fname=fname,
                         fname_print=fname_print,
                         line_no=LineNumber(line_no),
@@ -256,15 +339,15 @@ class ScaleneJSON:
                         profile_memory=profile_memory,
                         force_print=False,
                     )
-                    # o["percent_cpu_time"] = percent_cpu_time
-                    # o["elapsed_time"] = stats.elapsed_time
                     # Only output if the payload for the line is non-zero.
-                    if o:
-                        o_copy = copy.copy(o)
-                        del o_copy["line"]
-                        del o_copy["lineno"]
-                        if any(o_copy.values()):
-                            output["files"][fname_print]["lines"].append(o)
+                    if profile_line:
+                        profile_line_copy = copy.copy(profile_line)
+                        del profile_line_copy["line"]
+                        del profile_line_copy["lineno"]
+                        if any(profile_line_copy.values()):
+                            output["files"][fname_print]["lines"].append(
+                                profile_line
+                            )
             fn_stats = stats.build_function_stats(fname)
             # Check CPU samples and memory samples.
             print_fn_summary = False
@@ -277,12 +360,12 @@ class ScaleneJSON:
             output["files"][fname_print]["functions"] = []
             if print_fn_summary:
                 for fn_name in sorted(
-                        all_samples,
-                        key=lambda k: stats.firstline_map[k],
+                    all_samples,
+                    key=lambda k: stats.firstline_map[k],
                 ):
                     if fn_name == fname:
                         continue
-                    o = self.output_profile_line(
+                    profile_line = self.output_profile_line(
                         fname=fn_name,
                         fname_print=fn_name,
                         # line 1 is where function stats are
@@ -294,11 +377,13 @@ class ScaleneJSON:
                         profile_memory=profile_memory,
                         force_print=True,
                     )
-                    if o:
+                    if profile_line:
                         # Change the source code to just the function name.
-                        o["line"] = fn_name
+                        profile_line["line"] = fn_name
                         # Fix the line number to point to the first line of the function.
-                        o["lineno"] = stats.firstline_map[fn_name]
-                        output["files"][fname_print]["functions"].append(o)
+                        profile_line["lineno"] = stats.firstline_map[fn_name]
+                        output["files"][fname_print]["functions"].append(
+                            profile_line
+                        )
 
         return output
