@@ -128,7 +128,7 @@ for the process ID that Scalene reports. For example:
             action="store_const",
             const=True,
             default=defaults.cli,
-            help="forces use of the command-line"
+            help="forces use of the command-line",
         )
         parser.add_argument(
             "--web",
@@ -136,14 +136,14 @@ for the process ID that Scalene reports. For example:
             action="store_const",
             const=True,
             default=defaults.web,
-            help="writes 'profile.json' and opens the web UI (http://plasma-umass.org/scalene-gui/)"
+            help="writes 'profile.json' and opens the web UI (http://plasma-umass.org/scalene-gui/)",
         )
         parser.add_argument(
             "--port",
             dest="port",
             type=int,
             default=defaults.port,
-            help=f"binds the web UI server to this port (default: {defaults.port})"
+            help=f"binds the web UI server to this port (default: {defaults.port})",
         )
         parser.add_argument(
             "--reduced-profile",
