@@ -1,6 +1,6 @@
 import signal
 import sys
-from typing import List
+from typing import List, Tuple
 
 
 class ScaleneSignals:
@@ -37,6 +37,9 @@ class ScaleneSignals:
             else:
                 self.cpu_timer_signal = signal.ITIMER_REAL
                 self.cpu_signal = signal.SIGALRM
+
+    def get_timer_signals(self) -> Tuple[int, signal.Signals]:
+        return self.cpu_timer_signal, self.cpu_signal
 
     def get_all_signals(self) -> List[int]:
         return [
