@@ -1,1 +1,3 @@
-scalene_version = "1.5.3"
+"""Current version of Scalene; reported by --version."""
+
+scalene_version = "1.5.5"
