@@ -24,6 +24,7 @@
 // We're unable to use the limited API because, for example,
 // there doesn't seem to be a function returning co_filename
 //#define Py_LIMITED_API 0x03070000
+#include <execinfo.h>
 
 #include "common.hpp"
 #include "mallocrecursionguard.hpp"
@@ -42,7 +43,6 @@ static SampleFile& getSampleFile() {
 }
 
 #define USE_ATOMICS 0
-
 #if USE_ATOMICS
 typedef std::atomic<uint64_t> counterType;
 #else
@@ -100,11 +100,62 @@ class SampleHeap : public SuperHeap {
     }
     return ptr;
   }
+  ATTRIBUTE_ALWAYS_INLINE inline void* realloc(void* ptr, size_t sz) {
+    MallocRecursionGuard g;
+    if (!ptr) {
+      ptr = SuperHeap::malloc(sz);
+      return ptr;
+    }
+    if (sz == 0) {
+      SuperHeap::free(ptr);
+#if defined(__APPLE__)
+      // 0 size = free. We return a small object.  This behavior is
+      // apparently required under Mac OS X and optional under POSIX.
+      return SuperHeap::malloc(1);
+#else
+      // For POSIX, don't return anything.
+      return nullptr;
+#endif
+    }
 
+    size_t objSize = SuperHeap::getSize(ptr);
+
+    void* buf = SuperHeap::malloc(sz);
+    size_t buf_size = buf ? SuperHeap::getSize(buf) : 0;
+    if (buf) {
+      if (objSize == buf_size) {
+        // The objects are the same actual size.
+        // Free the new object and return the original.
+        SuperHeap::free(buf);
+        return ptr;
+      }
+      // Copy the contents of the original object
+      // up to the size of the new block.
+      size_t minSize = (objSize < sz) ? objSize : sz;
+      memcpy(buf, ptr, minSize);
+    }
+
+    // Free the old block.
+    SuperHeap::free(ptr);
+    if (buf) {
+      if (sz < buf_size) {
+        if (buf_size - sz <= PYMALLOC_MAX_SIZE) {
+          register_malloc(buf_size - sz, buf);
+        }
+      } else if (sz > buf_size) {
+        if (sz - buf_size <= PYMALLOC_MAX_SIZE) {
+          register_free(sz - buf_size, ptr);
+        }
+      }
+    }
+    // Return a pointer to the new one.
+    return buf;
+  }
   inline void register_malloc(size_t realSize, void* ptr,
                               bool inPythonAllocator = true) {
     assert(realSize);
     // If this is the special NEWLINE value, trigger an update.
+    // printf("uwu %p\n", &in_realloc);
     if (unlikely(realSize == NEWLINE)) {
       std::string filename;
       int lineno;
@@ -152,19 +203,19 @@ class SampleHeap : public SuperHeap {
  public:
   ATTRIBUTE_ALWAYS_INLINE inline void free(void* ptr) {
     MallocRecursionGuard g;
-
     if (unlikely(ptr == nullptr)) {
       return;
     }
+    auto realSize = SuperHeap::getSize(ptr);
+    SuperHeap::free(ptr);
     if (pythonDetected() && !g.wasInMalloc()) {
-      auto realSize = SuperHeap::getSize(ptr);
       register_free(realSize, ptr);
     }
-    SuperHeap::free(ptr);
   }
 
   inline void register_free(size_t realSize, void* ptr) {
     auto sampleFree = _allocationSampler.decrement(realSize);
+
     if (unlikely(ptr && (ptr == _lastMallocTrigger))) {
       _freedLastMallocTrigger = true;
     }
