@@ -1,35 +1,39 @@
 import tensorflow as tf
 from time import perf_counter
 
-num_threads = 16
-tf.config.threading.set_inter_op_parallelism_threads(
-    num_threads
-)
-tf.config.threading.set_intra_op_parallelism_threads(
-    num_threads
-)
+def config():
+    num_threads = 16
+    tf.config.threading.set_inter_op_parallelism_threads(
+        num_threads
+    )
+    tf.config.threading.set_intra_op_parallelism_threads(
+        num_threads
+    )
 
-mnist = tf.keras.datasets.mnist
+def run_benchmark():
+    mnist = tf.keras.datasets.mnist
 
-(x_train, y_train), (x_test, y_test) = mnist.load_data()
-x_train, x_test = x_train / 255.0, x_test / 255.0
+    (x_train, y_train), (x_test, y_test) = mnist.load_data()
+    x_train, x_test = x_train / 255.0, x_test / 255.0
 
-model = tf.keras.models.Sequential([
-  tf.keras.layers.Flatten(input_shape=(28, 28)),
-  tf.keras.layers.Dense(128, activation='relu'),
-  tf.keras.layers.Dropout(0.2),
-  tf.keras.layers.Dense(10)
-])
+    model = tf.keras.models.Sequential([
+      tf.keras.layers.Flatten(input_shape=(28, 28)),
+      tf.keras.layers.Dense(128, activation='relu'),
+      tf.keras.layers.Dropout(0.2),
+      tf.keras.layers.Dense(10)
+    ])
 
-predictions = model(x_train[:1]).numpy()
-print("predictions", predictions)
+    predictions = model(x_train[:1]).numpy()
+    print("predictions", predictions)
 
-loss_fn = tf.keras.losses.SparseCategoricalCrossentropy(from_logits=True)
-model.compile(optimizer='adam',
-              loss=loss_fn,
-              metrics=['accuracy'])
-t0 = perf_counter()
-model.fit(x_train, y_train, epochs=5)
-model.evaluate(x_test,  y_test, verbose=2)
-dt = perf_counter() - t0
-print(f"Total time: {dt}")
+    loss_fn = tf.keras.losses.SparseCategoricalCrossentropy(from_logits=True)
+    model.compile(optimizer='adam',
+                  loss=loss_fn,
+                  metrics=['accuracy'])
+    t0 = perf_counter()
+    model.fit(x_train, y_train, epochs=5)
+    model.evaluate(x_test,  y_test, verbose=2)
+    dt = perf_counter() - t0
+    print(f"Total time: {dt}")
+
+run_benchmark()
